refactor: App 컴포넌트 Page 폴더로 이동
