+++ src/pages/App/SideNavLink/index.tsx
@@ -1,10 +1,9 @@
 import route from '@Utils/route';
 
-import NavigationLink from '@Components/NavigationLink';
-
 import type { MainCategory } from '@Types/route';
 
 import * as S from './style';
+import NavigationLink from '../NavigationLink';
 
 function SideNavLink({ pathname }: { pathname: string }) {
   const [mainPath, subPath] = route.getPathnames(pathname);
